requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pyotp==2.9.0
qrcode==7.4.2
pillow==10.2.0
pyjwt==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
